partialise: calculate partial components

Calculate the partial components predicted by a model

Usage

partialise(obj)

# S3 method for class 'matrix'
partialise(obj, cdata, fml)

# S3 method for class 'DESeqDataSet'
partialise(obj, assay = "vst")

Arguments

obj

A DESeqDataSet object, where the design formula and colData will be used

cdata

A data.frame of metadata (rows=samples, columns=variables)

fml

A one-sided formula to predict each row of obj given the metadata

Value

A 'partialised' object

Details

Given a matrix-like object, calculate the amount that each term in a design formula contributes to the outcome. This, along with each sample's residual error, allows us to reconstruct the original matrix, in a customised way that includes or removes the contribution of specific variables.

See also

Author

Gavin Kelly