Generate captioned plot

quarto-compatible plot wrapper

Usage

plot_tracker(p)

Arguments

pl

A plot object, from ggplot, plot or ComplexHeatmap

label

The candidate chunk label, which will get sanitised

caption

The caption text

cap_fn

The function that will be called on the caption text

Details

Either produce a child chunk (labelled so that it can be cross-references), or open the default graphics device, depending on whether the document is being rendered or run interactively.

Author

Gavin Kelly