Recode nested factors

Recode nested factors to avoid matrix-rank problems

Usage

recode_within(inner, ...)

Arguments

inner

A factor representing the variable to be recoded, e.g. the cell-line

...

The parent factors that 'inner' is inside, e.g. the genotype of the cell-line

Value

A factor with recoded levels

Details

Following the suggestion in the DESeq2 vignette, recode nested factors so that they take common values in different clusters to avoid rank problems

Author

Gavin Kelly